+++ src/features/mod.rs
@@ -17,7 +17,7 @@ pub use beyond_n_std::BeyondNStd;
 
 pub(crate) mod bins;
 pub use bins::Bins;
-mod bootstrap;
+pub(crate) mod bootstrap;
 pub use bootstrap::{Bootstrap, BootstrapUncertainty};
 mod chi2_pvar;
 pub use chi2_pvar::Chi2Pvar;